Add information to tacos.html file
Add image and alt text of a taco for visual flair.
This will help the webpage be less bland.

Add description to taco recipe to describe the recipe.
Tell visitors about the recipe so they have an idea
what they will be eating.

Add list of ingredients describing what will be used in the dish.
Let the visitors know the ingredients they will need.

Add steps to follow when making the dish. Create a step by step
guide to making the recipe.